Split DelegationIRVisitor in multiple files
