--- src/PowerFx.Dataverse/Eval/Delegation/DelegationVisitor.cs
@@ -1,1630 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text.Json;
-using System.Text.Json.Serialization;
-using Microsoft.PowerFx.Core.IR;
-using Microsoft.PowerFx.Core.IR.Nodes;
-using Microsoft.PowerFx.Core.IR.Symbols;
-using Microsoft.PowerFx.Core.Localization;
-using Microsoft.PowerFx.Core.Texl;
-using Microsoft.PowerFx.Core.Types.Enums;
-using Microsoft.PowerFx.Core.Utils;
-using Microsoft.PowerFx.Dataverse.Eval.Core;
-using Microsoft.PowerFx.Dataverse.Eval.Delegation;
-using Microsoft.PowerFx.Dataverse.Eval.Delegation.DelegatedFunctions;
-using Microsoft.PowerFx.Types;
-using Microsoft.Xrm.Sdk.Metadata;
-using static Microsoft.PowerFx.Dataverse.DelegationEngineExtensions;
-using BinaryOpNode = Microsoft.PowerFx.Core.IR.Nodes.BinaryOpNode;
-using CallNode = Microsoft.PowerFx.Core.IR.Nodes.CallNode;
-using RecordNode = Microsoft.PowerFx.Core.IR.Nodes.RecordNode;
-using Span = Microsoft.PowerFx.Syntax.Span;
-using UnaryOpNode = Microsoft.PowerFx.Core.IR.Nodes.UnaryOpNode;
-
-namespace Microsoft.PowerFx.Dataverse
-{
-    // Rewrite the tree to inject delegation.
-    // If we encounter a dataverse table (something that should be delegated) during the walk, we either:
-    // - successfully delegate, which means rewriting to a call an efficient DelegatedFunction,
-    // - leave IR unchanged (don't delegate), but issue a warning.
-    internal class DelegationIRVisitor : RewritingIRVisitor<DelegationIRVisitor.RetVal, DelegationIRVisitor.Context>
-    {
-        // Ideally, this would just be in Dataverse.Eval nuget, but
-        // Only Dataverse nuget has InternalsVisisble access to implement an IR walker.
-        // So implement the walker in lower layer, and have callbacks into Dataverse.Eval layer as needed.
-        private readonly DelegationHooks _hooks;
-
-        private readonly int _maxRows;
-
-        // For reporting delegation Warnings.
-        private readonly ICollection<ExpressionError> _errors;
-
-        public DelegationIRVisitor(DelegationHooks hooks, ICollection<ExpressionError> errors, int maxRow)
-        {
-            _hooks = hooks ?? throw new ArgumentNullException(nameof(hooks));
-            _errors = errors ?? throw new ArgumentNullException(nameof(errors));
-            _maxRows = maxRow;
-        }
-
-        internal static readonly JsonSerializerOptions _options = new JsonSerializerOptions() { DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingDefault };
-
-        // Return Value passed through at each phase of the walk.
-        public class RetVal
-        {
-            private readonly IntermediateNode _filter;
-
-            private readonly IntermediateNode _orderBy;
-
-            private readonly IntermediateNode _topCount;
-
-            private readonly NumberLiteralNode _maxRows;
-
-            public bool HasFilter => _filter != null;
-
-            public bool HasOrderBy => _orderBy != null;
-
-            public bool HasTopCount => _topCount != null;
-
-            public bool HasColumnMap => ColumnMap != null;
-
-            public IntermediateNode Filter => _filter ?? Hooks.MakeBlankCall();
-
-            public IntermediateNode OrderBy => _orderBy ?? Hooks.MakeBlankCall();
-
-            public IntermediateNode TopCountOrDefault => _topCount ?? _maxRows;
-
-            // supports $select with column renames and distinct
-            public readonly ColumnMap ColumnMap;
-
-            public readonly DelegationHooks Hooks;
-
-            // Original IR node for non-delegating path.
-            // This should always be set. Even if we are attempting to delegate, we may need to use this if we can't support the delegation.
-            public readonly IntermediateNode OriginalNode;
-
-            // If set, we're attempting to delegate the current expression specifeid by _node.
-            public bool IsDelegating { get; init; }
-
-            // IR node that will resolve to the TableValue at runtime.
-            // From here, we can downcast at get the services. Ideally would be either Scope node or ResolvedObjectNode
-            public readonly DelegableIntermediateNode _sourceTableIRNode;
-
-            // Table type  and original metadata for table that we're delegating to.
-            public readonly TableType TableType;
-
-            // Null if not dataverse
-            private readonly EntityMetadata _metadata;
-
-            public bool TryGetPrimaryIdFieldName(out string primaryId)
-            {
-                if (_metadata == null)
-                {
-                    primaryId = null;
-                    return false;
-                }
-
-                primaryId = _metadata.PrimaryIdAttribute;
-                return true;
-            }
-
-            public bool TryGetLogicalName(out string logicalName)
-            {
-                if (_metadata != null)
-                {
-                    logicalName = _metadata.LogicalName;
-                    return true;
-                }
-
-                logicalName = null;
-                return false;
-            }
-
-            public bool IsElasticTable => _metadata?.IsElasticTable() ?? false;
-
-            public RetVal(DelegationHooks hooks, IntermediateNode originalNode, IntermediateNode sourceTableIRNode, TableType tableType, IntermediateNode filter, IntermediateNode orderBy, IntermediateNode count, int maxRows, ColumnMap columnMap)
-            {
-                this._maxRows = new NumberLiteralNode(IRContext.NotInSource(FormulaType.Number), maxRows);
-                this._sourceTableIRNode = new DelegableIntermediateNode(sourceTableIRNode ?? throw new ArgumentNullException(nameof(sourceTableIRNode)));
-                this.TableType = tableType ?? throw new ArgumentNullException(nameof(tableType));
-                this.OriginalNode = originalNode ?? throw new ArgumentNullException(nameof(originalNode));
-                this.Hooks = hooks ?? throw new ArgumentNullException(nameof(hooks));
-
-                // topCount and filter are optional.
-                this._topCount = count;
-                this._filter = filter;
-                this._orderBy = orderBy;
-                this.ColumnMap = columnMap;
-                this.IsDelegating = true;
-
-                _ = DelegationUtility.TryGetEntityMetadata(tableType, out this._metadata);
-            }
-
-            // Non-delegating path
-            public RetVal(IntermediateNode node)
-            {
-                OriginalNode = node;
-            }
-
-            internal IntermediateNode AddFilter(IntermediateNode newFilter, ScopeSymbol scope)
-            {
-                if (_filter != null)
-                {
-                    var combinedFilter = new List<IntermediateNode> { _filter, newFilter };
-                    var result = Hooks.MakeAndCall(TableType, combinedFilter, scope);
-                    return result;
-                }
-
-                return newFilter;
-            }
-        }
-
-        public class Context
-        {
-            public bool _ignoreDelegation;
-
-            public bool IsPredicateEvalInProgress => CallerNode != null && CallerTableNode != null;
-
-            public readonly CallNode CallerNode;
-
-            public readonly DelegableIntermediateNode CallerTableNode;
-
-            public readonly Stack<IDictionary<string, RetVal>> WithScopes;
-
-            public readonly RetVal CallerTableRetVal;
-
-            public Context()
-            {
-                WithScopes = new ();
-            }
-
-            private Context(bool ignoreDelegation, Stack<IDictionary<string, RetVal>> withScopes, CallNode callerNode, RetVal callerTableRetVal)
-            {
-                WithScopes = withScopes;
-                CallerNode = callerNode;
-                CallerTableNode = callerTableRetVal._sourceTableIRNode;
-                CallerTableRetVal = callerTableRetVal;
-                _ignoreDelegation = ignoreDelegation;
-            }
-
-            public Context GetContextForPredicateEval(CallNode callerNode, RetVal callerTableNode)
-            {
-                return new Context(this._ignoreDelegation, this.WithScopes, callerNode, callerTableNode);
-            }
-
-            internal void PushWithScope(IDictionary<string, RetVal> withScope)
-            {
-                WithScopes.Push(withScope);
-            }
-
-            internal IDictionary<string, RetVal> PopWithScope()
-            {
-                return WithScopes.Pop();
-            }
-        }
-
-        // If an attempted delegation can't be complete, then fail it.
-        private void AddError(ExpressionError error)
-        {
-            _errors.Add(error);
-        }
-
-        // Does this match: primaryKey=value
-        private bool MatchPrimaryId(IntermediateNode primaryIdField, RetVal tableArg)
-        {
-            return primaryIdField is ScopeAccessNode san && san.Value is ScopeAccessSymbol sas
-                ? tableArg.TryGetPrimaryIdFieldName(out var primaryId) && (sas.Name == primaryId)
-                : false;
-        }
-
-        // Normalize order? (Id=Guid) vs (Guid=Id)
-        private bool TryMatchPrimaryId(IntermediateNode left, IntermediateNode right, out IntermediateNode primaryIdField, out IntermediateNode guidValue, RetVal tableArg)
-        {
-            if (MatchPrimaryId(left, tableArg))
-            {
-                primaryIdField = left;
-                guidValue = right;
-                return true;
-            }
-            else if (MatchPrimaryId(right, tableArg))
-            {
-                primaryIdField = right;
-                guidValue = left;
-                return true;
-            }
-
-            primaryIdField = null;
-            guidValue = null;
-            return false;
-        }
-
-        // Issue warning on typo:
-        //  Filter(table, id=id)
-        //  LookUp(table, id=id)
-        //
-        // It's legal (so must be warning, not error). Likely, correct behavior is:
-        //  LookUp(table, ThisRecord.id=[@id])
-        private void CheckForNopLookup(CallNode node)
-        {
-            var func = node.Function.Name;
-            if (func == "LookUp" || func == "Filter")
-            {
-                if (node.Args.Count == 2)
-                {
-                    if (node.Args[1] is LazyEvalNode arg1b && arg1b.Child is BinaryOpNode predicate)
-                    {
-                        var left = predicate.Left;
-                        var right = predicate.Right;
-
-                        if (left is ScopeAccessNode left1 && right is ScopeAccessNode right1)
-                        {
-                            if (left1.Value is ScopeAccessSymbol left2 && right1.Value is ScopeAccessSymbol right2)
-                            {
-                                if (left2.Parent.Id == right2.Parent.Id &&
-                                    left2.Name == right2.Name)
-                                {
-                                    var reason = new ExpressionError()
-                                    {
-                                        Span = predicate.IRContext.SourceContext,
-                                        Severity = ErrorSeverity.Warning,
-                                        ResourceKey = TexlStrings.WrnDelegationPredicate
-                                    };
-                                    this.AddError(reason);
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        public override IntermediateNode Materialize(RetVal ret)
-        {
-            // if ret has no filter or count, then we can just return the original node.
-            if (ret.IsDelegating && (ret.HasFilter || ret.HasTopCount || ret.HasOrderBy || ret.HasColumnMap))
-            {
-                var res = _hooks.MakeQueryExecutorCall(ret);
-                return res;
-            }
-
-            return ret.OriginalNode;
-        }
-
-        protected override RetVal Ret(IntermediateNode node)
-        {
-            return new RetVal(node);
-        }
-
-        // ResolvedObject is a symbol injected by the host.
-        // All Table references start as resolved objects.
-        public override RetVal Visit(ResolvedObjectNode node, Context context)
-        {
-            if (!context._ignoreDelegation && node.IRContext.ResultType is TableType aggType)
-            {
-                // Does the resolve object refer to a dataverse Table?
-
-                if (node.Value is NameSymbol nameSym)
-                {
-                    var symbolTable = nameSym.Owner;
-
-                    // We need to tell the difference between a direct table,
-                    // and another global variable that has that table's type (such as global := Filter(table, true).
-                    bool isRealTable = _hooks.IsDelegableSymbolTable(symbolTable);
-
-                    if (isRealTable)
-                    {
-                        var ret = new RetVal(_hooks, node, node, aggType, filter: null, orderBy: null, count: null, _maxRows, columnMap: null);
-                        return ret;
-                    }
-                }
-            }
-
-            // Just a regular variable, don't bother delegating.
-            return Ret(node);
-        }
-
-        // BinaryOpNode can be only materialized when called via CallNode.
-        public override RetVal Visit(BinaryOpNode node, Context context)
-        {
-            if (!context.IsPredicateEvalInProgress)
-            {
-                return base.Visit(node, context);
-            }
-
-            var caller = context.CallerNode;
-            var callerReturnType = caller.IRContext.ResultType;
-            var callerSourceTable = context.CallerTableNode;
-            var callerScope = caller.Scope;
-
-            TableType tableType;
-            if (callerReturnType is RecordType recordType)
-            {
-                tableType = recordType.ToTable();
-            }
-            else
-            {
-                tableType = (TableType)callerReturnType;
-            }
-
-            IList<string> relations = null;
-
-            // Either left or right is field (not both)
-            if (!TryGetFieldName(context, node.Left, node.Right, node.Op, out var fieldName, out var rightNode, out var operation)
-                && !TryGetRelationField(context, node.Left, node.Right, node.Op, out fieldName, out relations, out rightNode, out operation))
-            {
-                return new RetVal(node);
-            }
-
-            if (context.CallerTableRetVal.HasColumnMap && context.CallerTableRetVal.ColumnMap.AsStringDictionary().TryGetValue(fieldName, out string realFieldName))
-            {
-                fieldName = realFieldName;
-            }
-
-            if (!IsRelationDelegationAllowed(tableType, relations))
-            {
-                return new RetVal(node);
-            }
-
-            var findThisRecord = ThisRecordIRVisitor.FindThisRecordUsage(caller, rightNode);
-            if (findThisRecord != null)
-            {
-                CreateThisRecordErrorAndReturn(caller, findThisRecord);
-            }
-
-            var findBehaviorFunc = BehaviorIRVisitor.Find(rightNode);
-            if (findBehaviorFunc != null)
-            {
-                return CreateBehaviorErrorAndReturn(caller, findBehaviorFunc);
-            }
-
-            var retDelegationVisitor = rightNode.Accept(this, context);
-            if (retDelegationVisitor.IsDelegating)
-            {
-                rightNode = Materialize(retDelegationVisitor);
-            }
-            else
-            {
-                rightNode = retDelegationVisitor.OriginalNode;
-            }
-
-            RetVal ret;
-
-            if (IsOpKindEqualityComparison(operation))
-            {
-                var eqNode = _hooks.MakeEqCall(callerSourceTable, tableType, relations, fieldName, operation, rightNode, callerScope);
-                ret = CreateBinaryOpRetVal(context, node, eqNode);
-            }
-            else if (IsOpKindInequalityComparison(operation))
-            {
-                var neqNode = _hooks.MakeNeqCall(callerSourceTable, tableType, relations, fieldName, operation, rightNode, callerScope);
-                ret = CreateBinaryOpRetVal(context, node, neqNode);
-            }
-            else if (IsOpKindLessThanComparison(operation))
-            {
-                var ltNode = _hooks.MakeLtCall(callerSourceTable, tableType, relations, fieldName, operation, rightNode, callerScope);
-                ret = CreateBinaryOpRetVal(context, node, ltNode);
-            }
-            else if (IsOpKindLessThanEqualComparison(operation))
-            {
-                var leqNode = _hooks.MakeLeqCall(callerSourceTable, tableType, relations, fieldName, operation, rightNode, callerScope);
-                ret = CreateBinaryOpRetVal(context, node, leqNode);
-            }
-            else if (IsOpKindGreaterThanComparison(operation))
-            {
-                var gtNode = _hooks.MakeGtCall(callerSourceTable, tableType, relations, fieldName, operation, rightNode, callerScope);
-                ret = CreateBinaryOpRetVal(context, node, gtNode);
-            }
-            else if (IsOpKindGreaterThanEqalComparison(operation))
-            {
-                var geqNode = _hooks.MakeGeqCall(callerSourceTable, tableType, relations, fieldName, operation, rightNode, callerScope);
-                ret = CreateBinaryOpRetVal(context, node, geqNode);
-            }
-            else
-            {
-                ret = new RetVal(node);
-            }
-
-            return ret;
-        }
-
-        private bool IsRelationDelegationAllowed(TableType tableType, IList<string> relations)
-        {
-            // Assume we can delegate if we can't find the metadata. It is possible if the table was the output of ShowColumns().
-            var result = true;
-
-            // For Elastic tables, we can delegate only if the field is a direct field of the table and NOT if it is a relation. As elastic table currently doesn't support relations in delegation.
-            if (DelegationUtility.TryGetEntityMetadata(tableType, out var metadata))
-            {
-                result = !(metadata.IsElasticTable() && relations != null && relations.Count > 0);
-            }
-
-            return result;
-        }
-
-        internal static bool IsOpKindEqualityComparison(BinaryOpKind op)
-        {
-            return op == BinaryOpKind.EqBoolean ||
-                op == BinaryOpKind.EqCurrency ||
-                op == BinaryOpKind.EqDate ||
-                op == BinaryOpKind.EqDateTime ||
-                op == BinaryOpKind.EqDecimals ||
-                op == BinaryOpKind.EqGuid ||
-                op == BinaryOpKind.EqNumbers ||
-                op == BinaryOpKind.EqText ||
-                op == BinaryOpKind.EqTime ||
-                op == BinaryOpKind.EqOptionSetValue ||
-                op == BinaryOpKind.EqPolymorphic;
-        }
-
-        internal static bool IsOpKindInequalityComparison(BinaryOpKind op)
-        {
-            return op == BinaryOpKind.NeqBoolean ||
-                op == BinaryOpKind.NeqCurrency ||
-                op == BinaryOpKind.NeqDate ||
-                op == BinaryOpKind.NeqDateTime ||
-                op == BinaryOpKind.NeqDecimals ||
-                op == BinaryOpKind.NeqGuid ||
-                op == BinaryOpKind.NeqNumbers ||
-                op == BinaryOpKind.NeqText ||
-                op == BinaryOpKind.NeqTime ||
-                op == BinaryOpKind.NeqOptionSetValue ||
-                op == BinaryOpKind.NeqPolymorphic;
-        }
-
-        internal static bool IsOpKindLessThanComparison(BinaryOpKind op)
-        {
-            return op == BinaryOpKind.LtNumbers ||
-                op == BinaryOpKind.LtDecimals ||
-                op == BinaryOpKind.LtDateTime ||
-                op == BinaryOpKind.LtDate ||
-                op == BinaryOpKind.LtTime;
-        }
-
-        internal static bool IsOpKindLessThanEqualComparison(BinaryOpKind op)
-        {
-            return op == BinaryOpKind.LeqNumbers ||
-                op == BinaryOpKind.LeqDecimals ||
-                op == BinaryOpKind.LeqDateTime ||
-                op == BinaryOpKind.LeqDate ||
-                op == BinaryOpKind.LeqTime;
-        }
-
-        internal static bool IsOpKindGreaterThanComparison(BinaryOpKind op)
-        {
-            return op == BinaryOpKind.GtNumbers ||
-                op == BinaryOpKind.GtDecimals ||
-                op == BinaryOpKind.GtDateTime ||
-                op == BinaryOpKind.GtDate ||
-                op == BinaryOpKind.GtTime;
-        }
-
-        internal static bool IsOpKindGreaterThanEqalComparison(BinaryOpKind op)
-        {
-            return op == BinaryOpKind.GeqNumbers ||
-                op == BinaryOpKind.GeqDecimals ||
-                op == BinaryOpKind.GeqDateTime ||
-                op == BinaryOpKind.GeqDate ||
-                op == BinaryOpKind.GeqTime;
-        }
-
-        private RetVal CreateBinaryOpRetVal(Context context, IntermediateNode node, IntermediateNode eqNode)
-        {
-            var callerTable = context.CallerTableNode;
-            var callerTableReturnType = callerTable.IRContext.ResultType as TableType ?? throw new InvalidOperationException("CallerTable ReturnType should always be TableType");
-            return new RetVal(_hooks, node, callerTable, callerTableReturnType, eqNode, orderBy: null, count: null, _maxRows, columnMap: null);
-        }
-
-        public override RetVal Visit(LazyEvalNode node, Context context)
-        {
-            if (!context.IsPredicateEvalInProgress)
-            {
-                return base.Visit(node, context);
-            }
-
-            var child = node.Child.Accept(this, context);
-
-            if (child.IsDelegating)
-            {
-                return child;
-            }
-            else
-            {
-                if (!ReferenceEquals(child.OriginalNode, node.Child))
-                {
-                    node = new LazyEvalNode(node.IRContext, child.OriginalNode);
-                }
-
-                return Ret(node);
-            }
-        }
-
-        public override RetVal Visit(SingleColumnTableAccessNode node, Context context)
-        {
-            var maybeDelegatedfrom = Materialize(node.From.Accept(this, context));
-
-            if (!ReferenceEquals(node.From, maybeDelegatedfrom))
-            {
-                var delegatedSCTANode = new SingleColumnTableAccessNode(node.IRContext, maybeDelegatedfrom, node.Field);
-                return Ret(delegatedSCTANode);
-            }
-
-            return Ret(node);
-        }
-
-        public override RetVal Visit(CallNode node, Context context)
-        {
-            var funcName = node.Function.Name;
-
-            if (funcName == BuiltinFunctionsCore.And.Name && context.IsPredicateEvalInProgress)
-            {
-                return ProcessAnd(node, context);
-            }
-            else if (funcName == BuiltinFunctionsCore.Or.Name && context.IsPredicateEvalInProgress)
-            {
-                return ProcessOr(node, context);
-            }
-            else if (funcName == BuiltinFunctionsCore.IsBlank.Name && context.IsPredicateEvalInProgress)
-            {
-                return ProcessIsBlank(node, context);
-            }
-
-            // Some functions don't require delegation.
-            // Using a table diretly as arg0 here doesn't generate a warning.
-            if (funcName == BuiltinFunctionsCore.IsBlank.Name ||
-                funcName == BuiltinFunctionsCore.IsError.Name ||
-                funcName == "Patch" || funcName == "Collect")
-            {
-                RetVal arg0c = node.Args[0].Accept(this, context);
-
-                return base.Visit(node, context, arg0c);
-            }
-
-            if (node.Args.Count == 0)
-            {
-                // Delegated functions require arg0 is the table.
-                // So a 0-length args can't delegate.
-                return base.Visit(node, context);
-            }
-
-            // Since With supports scopes, it needs to be processed differently.
-            if (funcName == BuiltinFunctionsCore.With.Name)
-            {
-                return ProcessWith(node, context);
-            }
-
-            // Only below function fulfills assumption that first arg is Table
-            if (!(node.Function.ParamTypes.Length > 0 && node.Function.ParamTypes[0].IsTable))
-            {
-                return base.Visit(node, context);
-            }
-
-            RetVal tableArg;
-
-            // special casing Scope access for With()
-            if (node.Args[0] is ScopeAccessNode scopedFirstArg && scopedFirstArg.IRContext.ResultType is TableType && scopedFirstArg.Value is ScopeAccessSymbol scopedSymbol
-                && TryGetScopedVariable(context.WithScopes, scopedSymbol.Name, out var scopedNode))
-            {
-                tableArg = scopedNode;
-            }
-            else
-            {
-                tableArg = node.Args[0].Accept(this, context);
-            }
-
-            if (!tableArg.IsDelegating)
-            {
-                return base.Visit(node, context, tableArg);
-            }
-
-            RetVal ret = funcName switch
-            {
-                "Distinct" => ProcessDistinct(node, tableArg, context),
-                "Filter" => ProcessFilter(node, tableArg, context),
-                "First" => ProcessFirst(node, tableArg),
-                "FirstN" => ProcessFirstN(node, tableArg),
-                "LookUp" => ProcessLookUp(node, tableArg, context),
-                "Sort" or "SortByColumns" => ProcessSort(node, tableArg, context),
-                "ShowColumns" => ProcessShowColumns(node, tableArg),
-                "ForAll" => ProcessForAll(node, tableArg, context),
-                _ => ProcessOtherFunctions(node, tableArg)
-            };
-
-            return ret;
-        }
-
-        private RetVal ProcessDistinct(CallNode node, RetVal tableArg, Context context)
-        {
-            IntermediateNode filter = tableArg.HasFilter ? tableArg.Filter : null;
-            IntermediateNode orderBy = tableArg.HasOrderBy ? tableArg.OrderBy : null;
-            IntermediateNode count = tableArg.HasTopCount ? tableArg.TopCountOrDefault : null;
-
-            context = context.GetContextForPredicateEval(node, tableArg);
-
-            // Distinct can't be delegated if: Return type is not primitive, or if the field is not a direct field of the table.
-            if (count != null
-                || !TryGetFieldName(context, ((LazyEvalNode)node.Args[1]).Child, out var fieldName)
-                || !IsReturnTypePrimitive(node.IRContext.ResultType)
-                || DelegationUtility.IsElasticTable(tableArg.TableType))
-            {
-                var materializeTable = Materialize(tableArg);
-                if (!ReferenceEquals(node.Args[0], materializeTable))
-                {
-                    var delegatedDistinct = new CallNode(node.IRContext, node.Function, node.Scope, new List<IntermediateNode>() { materializeTable, node.Args[1] });
-                    return Ret(delegatedDistinct);
-                }
-
-                return Ret(node);
-            }
-
-            // let's create a single column map ("Value", fieldName) with a distinct on fieldName
-            ColumnMap map = new ColumnMap(fieldName);
-
-            // Combine with an existing map
-            map = ColumnMap.Combine(tableArg.ColumnMap, map);
-
-            // change to original node to current node and appends columnSet and Distinct.
-            var resultingTable = new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, filter, orderBy: orderBy, count, _maxRows, map);
-
-            return resultingTable;
-        }
-
-        // $$$ We should block this at Authoring time.
-        private static bool IsReturnTypePrimitive(FormulaType returnType)
-        {
-            if (returnType is TableType tableType)
-            {
-                var column = tableType.FieldNames.First();
-                if (tableType.GetFieldType(column)._type.IsPrimitive)
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        private RetVal ProcessIsBlank(CallNode node, Context context)
-        {
-            IList<string> relations = null;
-            if (TryGetFieldName(context, node, out var fieldName)
-                || TryGetRelationField(context, node, out fieldName, out relations))
-            {
-                var blankNode = new CallNode(IRContext.NotInSource(FormulaType.Blank), BuiltinFunctionsCore.Blank);
-
-                // BinaryOpKind doesn't matter for IsBlank because all value will be compared to null, so just use EqText.
-                var eqNode = _hooks.MakeEqCall(context.CallerTableNode, context.CallerTableNode.IRContext.ResultType, relations, fieldName, BinaryOpKind.EqText, blankNode, context.CallerNode.Scope);
-                var ret = CreateBinaryOpRetVal(context, node, eqNode);
-                return ret;
-            }
-
-            RetVal arg0c = node.Args[0].Accept(this, context);
-
-            return base.Visit(node, context, arg0c);
-        }
-
-        private RetVal ProcessOtherFunctions(CallNode node, RetVal tableArg)
-        {
-            var maybeDelegableTable = Materialize(tableArg);
-
-            // If TableArg was delegable, then replace it and no need to add warning. As expr like Concat(Filter(), expr) works fine.
-            if (!ReferenceEquals(node.Args[0], maybeDelegableTable))
-            {
-                var delegableArgs = new List<IntermediateNode>() { maybeDelegableTable };
-                delegableArgs.AddRange(node.Args.Skip(1));
-                CallNode delegableCallNode;
-                if (node.Scope != null)
-                {
-                    delegableCallNode = new CallNode(node.IRContext, node.Function, node.Scope, delegableArgs);
-                }
-                else
-                {
-                    delegableCallNode = new CallNode(node.IRContext, node.Function, delegableArgs);
-                }
-
-                return Ret(delegableCallNode);
-            }
-
-            return CreateNotSupportedErrorAndReturn(node, tableArg);
-        }
-
-        private RetVal ProcessShowColumns(CallNode node, RetVal tableArg)
-        {
-            IntermediateNode filter = tableArg.HasFilter ? tableArg.Filter : null;
-            IntermediateNode orderBy = tableArg.HasOrderBy ? tableArg.OrderBy : null;
-            IntermediateNode count = tableArg.HasTopCount ? tableArg.TopCountOrDefault : null;
-
-            // ShowColumns is only a column selector, so let's create a map with (column, column) entries
-            ColumnMap map = new ColumnMap(node.Args.Skip(1).Select(i => i is TextLiteralNode tln ? tln : throw new InvalidOperationException($"Expecting {nameof(TextLiteralNode)} and received {i.GetType().Name}")));
-
-            map = ColumnMap.Combine(tableArg.ColumnMap, map);
-
-            // change to original node to current node and appends columnSet.
-            var resultingTable = new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, filter, orderBy: orderBy, count, _maxRows, map);
-
-            if (node is CallNode maybeGuidCall && maybeGuidCall.Function is DelegatedRetrieveGUIDFunction)
-            {
-                var guidCallWithColSet = _hooks.MakeRetrieveCall(resultingTable, maybeGuidCall.Args[1]);
-                return Ret(guidCallWithColSet);
-            }
-
-            return resultingTable;
-        }
-
-        private RetVal ProcessWith(CallNode node, Context context)
-        {
-            IntermediateNode arg0Intermediate;
-            IntermediateNode arg1Intermediate;
-            var arg1 = (LazyEvalNode)node.Args[1];
-
-            if (node.Args[0] is RecordNode arg0)
-            {
-                var withScope = RecordNodeToDictionary(arg0, context);
-                arg0Intermediate = new RecordNode(arg0.IRContext, withScope.ToDictionary(kv => new DName(kv.Key), kv => Materialize(kv.Value)));
-
-                context.PushWithScope(withScope);
-                arg1Intermediate = Materialize(arg1.Child.Accept(this, context));
-                var poppedWithScope = context.PopWithScope();
-
-                if (withScope != poppedWithScope)
-                {
-                    throw new InvalidOperationException("With scope stack is corrupted");
-                }
-            }
-            else
-            {
-                arg0Intermediate = Materialize(node.Args[0].Accept(this, context));
-                arg1Intermediate = Materialize(arg1.Child.Accept(this, context));
-            }
-
-            if (!ReferenceEquals(arg1Intermediate, arg1.Child))
-            {
-                var lazyArg1 = new LazyEvalNode(arg1.Child.IRContext, arg1Intermediate);
-                return Ret(new CallNode(node.IRContext, node.Function, node.Scope, new List<IntermediateNode>() { arg0Intermediate, lazyArg1 }));
-            }
-            else
-            {
-                return Ret(new CallNode(node.IRContext, node.Function, node.Scope, new List<IntermediateNode>() { arg0Intermediate, arg1 }));
-            }
-        }
-
-        private IDictionary<string, RetVal> RecordNodeToDictionary(RecordNode arg0, Context context)
-        {
-            var scope = new Dictionary<string, RetVal>();
-            foreach (var field in arg0.Fields)
-            {
-                var valueRetVal = field.Value.Accept(this, context);
-                scope.Add(field.Key.Value, valueRetVal);
-            }
-
-            return scope;
-        }
-
-        private bool TryGetScopedVariable(Stack<IDictionary<string, RetVal>> withScopes, string variable, out RetVal node)
-        {
-            if (withScopes.Count() == 0)
-            {
-                node = default;
-                return false;
-            }
-
-            foreach (var kv in withScopes)
-            {
-                if (kv.TryGetValue(variable, out node))
-                {
-                    return true;
-                }
-            }
-
-            node = default;
-            return false;
-        }
-
-        private RetVal ProcessLookUp(CallNode node, RetVal tableArg, Context context)
-        {
-            RetVal result;
-            if (node.Args.Count != 2)
-            {
-                return CreateNotSupportedErrorAndReturn(node, tableArg);
-            }
-
-            IntermediateNode orderBy = tableArg.HasOrderBy ? tableArg.OrderBy : null;
-
-            var predicate = node.Args[1];
-            var predicteContext = context.GetContextForPredicateEval(node, tableArg);
-
-            // Pattern match to see if predicate is GUID delegable.
-            if (predicate is LazyEvalNode arg1b)
-            {
-                if (arg1b.Child is BinaryOpNode binOp &&
-                    binOp.Op == BinaryOpKind.EqGuid &&
-                    TryMatchPrimaryId(binOp.Left, binOp.Right, out _, out var guidValue, tableArg))
-                {
-                    CheckForNopLookup(node);
-
-                    // Pattern match to see if predicate is delegable.
-                    //  Lookup(Table, Id=Guid)
-                    var retVal2 = guidValue.Accept(this, context);
-                    var right = Materialize(retVal2);
-
-                    var findThisRecord = ThisRecordIRVisitor.FindThisRecordUsage(node, right);
-                    if (findThisRecord == null)
-                    {
-                        var findBehaviorFunc = BehaviorIRVisitor.Find(right);
-                        if (findBehaviorFunc != null)
-                        {
-                            CreateBehaviorErrorAndReturn(node, findBehaviorFunc);
-                        }
-                        else
-                        {
-                            // We can successfully delegate this call.
-                            // __retrieveGUID(table, guid);
-
-                            if (IsTableArgLookUpDelegable(context, tableArg))
-                            {
-                                var newNode = _hooks.MakeRetrieveCall(tableArg, right);
-                                return Ret(newNode);
-                            }
-                            else
-                            {
-                                // if tableArg was another delegation (e.g. Filter()), then we need to Materialize the call and can't delegate lookup.
-                                return MaterializeTableAndAddWarning(tableArg, node);
-                            }
-                        }
-                    }
-                    else
-                    {
-                        return CreateThisRecordErrorAndReturn(node, findThisRecord);
-                    }
-                }
-                else if (arg1b.Child is CallNode arg1MaybeAnd && arg1MaybeAnd.Function.Name == BuiltinFunctionsCore.And.Name && arg1MaybeAnd.Args.Count == 2)
-                {
-                    // If LookUp predicate only has primary key comparison and partition id and table is elastic table, then we can delegate the call.
-                    var arg1OfAnd = arg1MaybeAnd.Args[0];
-                    var arg2OfAnd = ((LazyEvalNode)arg1MaybeAnd.Args[1]).Child;
-                    if (TryMatchElasticIds(predicteContext, arg1OfAnd, arg2OfAnd, out var guidArg, out var partitionIdArg))
-                    {
-                        var guidRetVal = guidArg.Accept(this, context);
-                        var partitionIdRetVal = partitionIdArg.Accept(this, context);
-                        var materializedGuid = Materialize(guidRetVal);
-                        var materializedPartitionId = Materialize(partitionIdRetVal);
-                        if (IsTableArgLookUpDelegable(context, tableArg))
-                        {
-                            var newNode = _hooks.MakeElasticRetrieveCall(tableArg, materializedGuid, materializedPartitionId);
-                            return Ret(newNode);
-                        }
-                        else
-                        {
-                            // if tableArg was another delegation (e.g. Filter()), then we need to Materialize the call and can't delegate lookup.
-                            return MaterializeTableAndAddWarning(tableArg, node);
-                        }
-                    }
-                }
-            }
-
-            // Pattern match to see if predicate is delegable when field is non primary key.
-            var pr = predicate.Accept(this, predicteContext);
-
-            if (!pr.IsDelegating)
-            {
-                // Though entire predicate is not delegable, pr._originalNode may still have delegation buried inside it.
-                if (!ReferenceEquals(pr.OriginalNode, predicate))
-                {
-                    node = new CallNode(node.IRContext, node.Function, node.Scope, new List<IntermediateNode>() { node.Args[0], pr.OriginalNode });
-                }
-
-                return CreateNotSupportedErrorAndReturn(node, tableArg);
-            }
-
-            // if tableArg was DV Table, delegate the call.
-            if (IsTableArgLookUpDelegable(context, tableArg))
-            {
-                var filterCombined = tableArg.AddFilter(pr.Filter, node.Scope);
-                result = new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, filterCombined, orderBy: orderBy, count: null, _maxRows, tableArg.ColumnMap);
-            }
-            else
-            {
-                // if tableArg was a other delegation (e.g. Filter()), then we need to Materialize the call and can't delegate lookup.
-                result = MaterializeTableAndAddWarning(tableArg, node);
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// If predicate is comparing primary key and partition id, and table is elastic table, then we can delegate the call to a faster api.
-        /// </summary>
-        /// <param name="predicteContext"></param>
-        /// <param name="arg1OfAnd"></param>
-        /// <param name="arg2OfAnd"></param>
-        /// <param name="guidArg"></param>
-        /// <param name="partitionIdArg"></param>
-        /// <returns></returns>
-        private bool TryMatchElasticIds(Context predicteContext, IntermediateNode arg1OfAnd, IntermediateNode arg2OfAnd, out IntermediateNode guidArg, out IntermediateNode partitionIdArg)
-        {
-            if (predicteContext.CallerTableRetVal.IsElasticTable && arg1OfAnd is BinaryOpNode arg1b && arg2OfAnd is BinaryOpNode arg2b)
-            {
-                if (TryMatchPrimaryId(arg1b.Left, arg1b.Right, out _, out guidArg, predicteContext.CallerTableRetVal)
-                    && TryGetFieldName(predicteContext, arg2b.Left, arg2b.Right, arg2b.Op, out var fieldName, out var maybePartitionId, out var opKind)
-                    && fieldName == "partitionid"
-                    && IsOpKindEqualityComparison(opKind))
-                {
-                    partitionIdArg = maybePartitionId;
-                    return true;
-                }
-                else if (TryMatchPrimaryId(arg2b.Left, arg2b.Right, out _, out guidArg, predicteContext.CallerTableRetVal)
-                    && TryGetFieldName(predicteContext, arg1b.Left, arg1b.Right, arg1b.Op, out fieldName, out maybePartitionId, out opKind)
-                    && fieldName == "partitionid"
-                    && IsOpKindEqualityComparison(opKind))
-                {
-                    partitionIdArg = maybePartitionId;
-                    return true;
-                }
-            }
-
-            guidArg = null;
-            partitionIdArg = null;
-            return false;
-        }
-
-        private bool IsTableArgLookUpDelegable(Context context, RetVal tableArg)
-        {
-            if (tableArg.OriginalNode is ResolvedObjectNode
-                || (tableArg._sourceTableIRNode.InnerNode is ScopeAccessNode scopedTableArg
-                    && scopedTableArg.Value is ScopeAccessSymbol scopedSymbol
-                    && TryGetScopedVariable(context.WithScopes, scopedSymbol.Name, out var scopedNode)
-                    && scopedNode.OriginalNode is ResolvedObjectNode)
-                || (tableArg.IsDelegating && tableArg.OriginalNode is CallNode callNode
-                    && (callNode.Function.Name == "ShowColumns" ||
-                        callNode.Function.Name == "Distinct" ||
-                        callNode.Function.Name == "Sort" ||
-                        callNode.Function.Name == "SortByColumns")))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        private RetVal ProcessFilter(CallNode node, RetVal tableArg, Context context)
-        {
-            if (node.Args.Count != 2)
-            {
-                return CreateNotSupportedErrorAndReturn(node, tableArg);
-            }
-
-            IntermediateNode predicate = node.Args[1];
-            IntermediateNode orderBy = tableArg.HasOrderBy ? tableArg.OrderBy : null;
-
-            var predicteContext = context.GetContextForPredicateEval(node, tableArg);
-            var pr = predicate.Accept(this, predicteContext);
-
-            if (!pr.IsDelegating)
-            {
-                // Though entire predicate is not delegable, pr._originalNode may still have delegation buried inside it.
-                if (!ReferenceEquals(pr.OriginalNode, predicate))
-                {
-                    node = new CallNode(node.IRContext, node.Function, node.Scope, new List<IntermediateNode>() { node.Args[0], pr.OriginalNode });
-                }
-
-                return CreateNotSupportedErrorAndReturn(node, tableArg);
-            }
-
-            RetVal result;
-
-            // If tableArg has top count, that means we need to materialize the tableArg and can't delegate.
-            if (tableArg.HasTopCount)
-            {
-                result = MaterializeTableAndAddWarning(tableArg, node);
-            }
-            else
-            {
-                // Since table was delegating it potentially has filter attached to it, so also add that filter to the new filter.
-                var filterCombined = tableArg.AddFilter(pr.Filter, node.Scope);
-                result = new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, filterCombined, orderBy: orderBy, count: null, _maxRows, tableArg.ColumnMap);
-            }
-
-            return result;
-        }
-
-        private RetVal ProcessSort(CallNode node, RetVal tableArg, Context context)
-        {
-            IntermediateNode filter = tableArg.HasFilter ? tableArg.Filter : null;
-            IntermediateNode count = tableArg.HasTopCount ? tableArg.TopCountOrDefault : null;
-
-            List<IntermediateNode> arguments = new List<IntermediateNode>() { filter ?? node.Args[0] };
-
-            context = context.GetContextForPredicateEval(node, tableArg);
-
-            // If existing First[N], Sort[ByColumns], or ShowColumns we don't delegate
-            // When multiple Sort would occur, we cannot reliably group OrderBy commands
-            if (tableArg.HasTopCount || tableArg.HasOrderBy || tableArg.HasColumnMap)
-            {
-                return NoTransform(node, tableArg);
-            }
-
-            int i = 1;
-
-            while (i < node.Args.Count)
-            {
-                string fieldName;
-
-                if (node.Args[i] is TextLiteralNode tln)
-                {
-                    fieldName = tln.LiteralValue;
-                }
-                else if (!TryGetFieldName(context, ((LazyEvalNode)node.Args[i]).Child, out fieldName))
-                {
-                    return NoTransform(node, tableArg);
-                }
-
-                arguments.Add(new TextLiteralNode(IRContext.NotInSource(FormulaType.String), fieldName));
-
-                i++;
-                bool isAscending = true;
-
-                // this argument is optional and will default to Ascending if not provided
-                if (i < node.Args.Count)
-                {
-                    if (TryGetEnumValue(node.Args[i], "SortOrder", out string sortOrder))
-                    {
-                        isAscending = sortOrder.Equals("Ascending", StringComparison.Ordinal);
-                        i++;
-                    }
-                    else
-                    {
-                        return NoTransform(node, tableArg);
-                    }
-                }
-
-                arguments.Add(new BooleanLiteralNode(IRContext.NotInSource(FormulaType.Boolean), isAscending));
-            }
-
-            var sortFunc = new DelegatedSort(_hooks);
-            IntermediateNode orderByNode = new CallNode(node.IRContext, sortFunc, arguments);
-
-            var resultingTable = new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, filter, orderBy: orderByNode, count, _maxRows, tableArg.ColumnMap);
-
-            return resultingTable;
-        }
-
-        private RetVal ProcessForAll(CallNode node, RetVal tableArg, Context context)
-        {
-            IntermediateNode filter = tableArg.HasFilter ? tableArg.Filter : null;
-            IntermediateNode count = tableArg.HasTopCount ? tableArg.TopCountOrDefault : null;
-            IntermediateNode orderBy = tableArg.HasOrderBy ? tableArg.OrderBy : null;
-
-            context = context.GetContextForPredicateEval(node, tableArg);
-
-            // check if we have a simple field name here
-            if (TryGetFieldName(context, ((LazyEvalNode)node.Args[1]).Child, out string fieldName))
-            {
-                TextLiteralNode column = new TextLiteralNode(IRContext.NotInSource(FormulaType.String), fieldName);
-
-                // Create a map with ("Value", fieldName)
-                ColumnMap map = new ColumnMap(new Dictionary<DName, TextLiteralNode>() { { new DName("Value"), column } });
-
-                // Combine with an existing map
-                map = ColumnMap.Combine(tableArg.ColumnMap, map);
-
-                return new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, filter, orderBy, count, _maxRows, map);
-            }
-
-            // check if we have a record of (newName: oldName)
-            if (((LazyEvalNode)node.Args[1]).Child is RecordNode recordNode)
-            {
-                Dictionary<DName, TextLiteralNode> dic = new Dictionary<DName, TextLiteralNode>();
-                bool canDelegate = true;
-
-                foreach (KeyValuePair<DName, IntermediateNode> kvp in recordNode.Fields)
-                {
-                    string newFieldName = kvp.Key.Value;
-
-                    if (TryGetFieldName(context, kvp.Value, out string currentFieldName))
-                    {
-                        TextLiteralNode currentColumn = new TextLiteralNode(IRContext.NotInSource(FormulaType.String), currentFieldName);
-                        dic.Add(new DName(newFieldName), currentColumn);
-                    }
-                    else
-                    {
-                        // If any record field is not a trivial field name, we'll not delegate
-                        // ex: { newName: oldName * 2 }
-                        canDelegate = false;
-                        break;
-                    }
-                }
-
-                if (canDelegate)
-                {
-                    // Combine with an existing map
-                    ColumnMap map = ColumnMap.Combine(tableArg.ColumnMap, new ColumnMap(dic));
-
-                    return new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, filter, orderBy, count, _maxRows, map);
-                }
-            }
-
-            return ProcessOtherFunctions(node, tableArg);
-        }
-
-        private RetVal NoTransform(CallNode node, RetVal tableArg)
-        {
-            IntermediateNode materializeTable = Materialize(tableArg);
-
-            if (!ReferenceEquals(node.Args[0], materializeTable))
-            {
-                List<IntermediateNode> arguments = new List<IntermediateNode>() { materializeTable };
-                arguments.AddRange(node.Args.Skip(1));
-
-                CallNode delegatedSort = new CallNode(node.IRContext, node.Function, node.Scope, arguments);
-                return Ret(delegatedSort);
-            }
-
-            return Ret(node);
-        }
-
-        private bool TryGetEnumValue(IntermediateNode arg, string enumName, out string val)
-        {
-            if (arg is RecordFieldAccessNode rfan && rfan.From is ResolvedObjectNode ron && ron.Value is EnumSymbol es && es.EntityName.Value == enumName)
-            {
-                val = rfan.Field.Value;
-                return true;
-            }
-
-            val = null;
-            return false;
-        }
-
-        private RetVal ProcessFirst(CallNode node, RetVal tableArg)
-        {
-            IntermediateNode orderBy = tableArg.HasOrderBy ? tableArg.OrderBy : null;
-
-            var countOne = new NumberLiteralNode(IRContext.NotInSource(FormulaType.Number), 1);
-            var res = new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, tableArg.Filter, orderBy: orderBy, countOne, _maxRows, tableArg.ColumnMap);
-            return res;
-        }
-
-        private RetVal ProcessFirstN(CallNode node, RetVal tableArg)
-        {
-            IntermediateNode orderBy = tableArg.HasOrderBy ? tableArg.OrderBy : null;
-
-            // Add default count of 1 if not specified.
-            if (node.Args.Count == 1)
-            {
-                node.Args.Add(new NumberLiteralNode(IRContext.NotInSource(FormulaType.Number), 1));
-            }
-            else if (node.Args.Count != 2)
-            {
-                return CreateNotSupportedErrorAndReturn(node, tableArg);
-            }
-
-            return new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, tableArg.Filter, orderBy: orderBy, node.Args[1], _maxRows, tableArg.ColumnMap);
-        }
-
-        private RetVal MaterializeTableAndAddWarning(RetVal tableArg, CallNode node)
-        {
-            var tableCallNode = Materialize(tableArg);
-            var args = new List<IntermediateNode>() { tableCallNode, node.Args[1] };
-            var newCall = _hooks.MakeCallNode(node.Function, node.IRContext, args, node.Scope);
-            return CreateNotSupportedErrorAndReturn(newCall, tableArg);
-        }
-
-        private RetVal CreateNotSupportedErrorAndReturn(CallNode node, RetVal tableArg)
-        {
-            if (tableArg == null || !tableArg.TryGetLogicalName(out var tableName))
-            {
-                tableName = "table"; // some default
-            }
-
-            var reason = new ExpressionError()
-            {
-                MessageArgs = new object[] { tableName, _maxRows },
-                Span = tableArg?._sourceTableIRNode.IRContext.SourceContext ?? new Span(1, 2),
-                Severity = ErrorSeverity.Warning,
-                ResourceKey = TexlStrings.WrnDelegationTableNotSupported
-            };
-
-            this.AddError(reason);
-
-            return new RetVal(node);
-        }
-
-        private RetVal CreateBehaviorErrorAndReturn(CallNode node, BehaviorIRVisitor.RetVal findBehaviorFunc)
-        {
-            var reason = new ExpressionError()
-            {
-                MessageArgs = new object[] { node.Function.Name, findBehaviorFunc.Name },
-                Span = findBehaviorFunc.Span,
-                Severity = ErrorSeverity.Warning,
-                ResourceKey = TexlStrings.WrnDelegationBehaviorFunction
-            };
-
-            AddError(reason);
-            return new RetVal(node);
-        }
-
-        private RetVal CreateThisRecordErrorAndReturn(CallNode node, ThisRecordIRVisitor.RetVal findThisRecord)
-        {
-            var reason = new ExpressionError()
-            {
-                MessageArgs = new object[] { node.Function.Name },
-                Span = findThisRecord.Span,
-                Severity = ErrorSeverity.Warning,
-                ResourceKey = TexlStrings.WrnDelegationRefersThisRecord
-            };
-            AddError(reason);
-            return new RetVal(node);
-        }
-
-        private RetVal ProcessAndOr(CallNode node, Context context, Func<IList<IntermediateNode>, IntermediateNode> filterDelegate)
-        {
-            bool isDelegating = true;
-            List<IntermediateNode> delegatedChild = new ();
-
-            foreach (var arg in node.Args)
-            {
-                var delegatedArg = arg is LazyEvalNode lazyEvalNode
-                ? lazyEvalNode.Child.Accept(this, context)
-                : arg.Accept(this, context);
-
-                if (delegatedArg.IsDelegating)
-                {
-                    if (delegatedArg.HasFilter)
-                    {
-                        delegatedChild.Add(delegatedArg.Filter);
-                    }
-                }
-                else
-                {
-                    isDelegating = false;
-                    break;
-                }
-            }
-
-            if (isDelegating)
-            {
-                var filter = filterDelegate(delegatedChild);
-                var rVal = CreateBinaryOpRetVal(context, node, filter);
-                return rVal;
-            }
-
-            return new RetVal(node);
-        }
-
-        public RetVal ProcessOr(CallNode node, Context context)
-        {
-            var callerReturnType = context.CallerNode.IRContext.ResultType;
-            return ProcessAndOr(node, context, delegatedChildren => _hooks.MakeOrCall(callerReturnType, delegatedChildren, node.Scope));
-        }
-
-        public RetVal ProcessAnd(CallNode node, Context context)
-        {
-            var callerReturnType = context.CallerNode.IRContext.ResultType;
-            return ProcessAndOr(node, context, delegatedChildren => _hooks.MakeAndCall(callerReturnType, delegatedChildren, node.Scope));
-        }
-
-        public bool TryGetFieldName(Context context, IntermediateNode left, IntermediateNode right, BinaryOpKind op, out string fieldName, out IntermediateNode node, out BinaryOpKind opKind)
-        {
-            if (TryGetFieldName(context, left, out var leftField) && !TryGetFieldName(context, right, out _))
-            {
-                fieldName = leftField;
-                node = right;
-                opKind = op;
-                return true;
-            }
-            else if (TryGetFieldName(context, right, out var rightField) && !TryGetFieldName(context, left, out _))
-            {
-                fieldName = rightField;
-                node = left;
-                if (TryInvertLeftRight(op, out var invertedOp))
-                {
-                    opKind = invertedOp;
-                    return true;
-                }
-                else
-                {
-                    opKind = default;
-                    return false;
-                }
-            }
-            else if (TryGetFieldName(context, left, out var leftField2) && TryGetFieldName(context, right, out var rightField2))
-            {
-                if (leftField2 == rightField2)
-                {
-                    // Issue warning
-                    if (IsOpKindEqualityComparison(op))
-                    {
-                        var min = left.IRContext.SourceContext.Lim;
-                        var lim = right.IRContext.SourceContext.Min;
-                        var span = new Span(min, lim);
-                        var reason = new ExpressionError()
-                        {
-                            Span = span,
-                            Severity = ErrorSeverity.Warning,
-                            ResourceKey = TexlStrings.WrnDelegationPredicate
-                        };
-                        this.AddError(reason);
-                    }
-
-                    opKind = op;
-                    fieldName = default;
-                    node = default;
-                    return false;
-                }
-            }
-
-            opKind = op;
-            node = default;
-            fieldName = default;
-            return false;
-        }
-
-        private bool TryGetRelationField(Context context, IntermediateNode left, IntermediateNode right, BinaryOpKind op, out string fieldName, out IList<string> relations, out IntermediateNode node, out BinaryOpKind opKind)
-        {
-            if (TryGetRelationField(context, left, out var leftField, out var leftRelation) && !TryGetFieldName(context, right, out _))
-            {
-                fieldName = leftField;
-                relations = leftRelation;
-                node = right;
-                opKind = op;
-                return true;
-            }
-            else if (TryGetRelationField(context, right, out var rightField, out var rightRelation) && !TryGetFieldName(context, left, out _))
-            {
-                fieldName = rightField;
-                relations = rightRelation;
-                node = left;
-                if (TryInvertLeftRight(op, out var invertedOp))
-                {
-                    opKind = invertedOp;
-                    return true;
-                }
-                else
-                {
-                    opKind = default;
-                    return false;
-                }
-            }
-            else if (TryGetRelationField(context, left, out var leftField2, out _) && TryGetRelationField(context, right, out var rightField2, out _))
-            {
-                if (leftField2 == rightField2)
-                {
-                    // Issue warning
-                    if (IsOpKindEqualityComparison(op))
-                    {
-                        var min = left.IRContext.SourceContext.Lim;
-                        var lim = right.IRContext.SourceContext.Min;
-                        var span = new Span(min, lim);
-                        var reason = new ExpressionError()
-                        {
-                            Span = span,
-                            Severity = ErrorSeverity.Warning,
-                            ResourceKey = TexlStrings.WrnDelegationPredicate
-                        };
-                        this.AddError(reason);
-                    }
-
-                    opKind = op;
-                    fieldName = default;
-                    relations = default;
-                    node = default;
-                    return false;
-                }
-            }
-
-            opKind = op;
-            node = default;
-            fieldName = default;
-            relations = default;
-            return false;
-        }
-
-        private bool TryGetRelationField(Context context, IntermediateNode node, out string fieldName, out IList<string> relations)
-        {
-            relations = new List<string>();
-
-            IntermediateNode maybeFieldAccessNode;
-
-            // If the node had injected float coercion, then we need to pull field access node from it.
-            if (node is CallNode functionCall &&
-                (functionCall.Function == BuiltinFunctionsCore.Float || functionCall.Function == BuiltinFunctionsCore.Value || functionCall.Function.Name == BuiltinFunctionsCore.IsBlank.Name) &&
-                functionCall.Args.Count == 1)
-            {
-                maybeFieldAccessNode = functionCall.Args[0];
-            }
-            else
-            {
-                maybeFieldAccessNode = node;
-            }
-
-            if (maybeFieldAccessNode is RecordFieldAccessNode fieldAccess)
-            {
-                fieldName = fieldAccess.Field;
-                if (TryGetFieldName(context, fieldAccess.From, out var fromField))
-                {
-                    var relationMetadata = new RelationMetadata(fromField, false, null);
-
-                    var serializedRelationMetadata = DelegationUtility.SerializeRelationMetadata(relationMetadata);
-                    relations.Add(serializedRelationMetadata);
-                    return true;
-                }
-                else if (fieldAccess.From is CallNode callNode && callNode.Function.Name == BuiltinFunctionsCore.AsType.Name)
-                {
-                    if (TryGetEntityName(callNode.Args[1].IRContext.ResultType, out var targetEntityName) && TryGetEntityName(context.CallerTableNode.IRContext.ResultType, out _))
-                    {
-                        TryGetFieldName(context, callNode.Args[0], out fromField);
-                        AttributeUtility.TryGetLogicalNameFromOdataName(fromField, out var logicalName);
-                        var relationMetadata = new RelationMetadata(logicalName, true, targetEntityName);
-
-                        var serializedRelationMetadata = DelegationUtility.SerializeRelationMetadata(relationMetadata);
-                        relations.Add(serializedRelationMetadata);
-                        return true;
-                    }
-                }
-            }
-
-            fieldName = default;
-            return false;
-        }
-
-        private static bool TryGetEntityName(FormulaType type, out string entityName)
-        {
-            var ads = type._type.AssociatedDataSources.FirstOrDefault();
-            if (ads != null)
-            {
-                entityName = ads.EntityName.Value;
-                return true;
-            }
-
-            entityName = default;
-            return false;
-        }
-
-        private bool TryInvertLeftRight(BinaryOpKind op, out BinaryOpKind invertedOp)
-        {
-            switch (op)
-            {
-                case BinaryOpKind.LtNumbers:
-                    invertedOp = BinaryOpKind.GtNumbers;
-                    return true;
-
-                case BinaryOpKind.LtDecimals:
-                    invertedOp = BinaryOpKind.GtDecimals;
-                    return true;
-
-                case BinaryOpKind.LtDateTime:
-                    invertedOp = BinaryOpKind.GtDateTime;
-                    return true;
-
-                case BinaryOpKind.LtDate:
-                    invertedOp = BinaryOpKind.GtDate;
-                    return true;
-
-                case BinaryOpKind.LtTime:
-                    invertedOp = BinaryOpKind.GtTime;
-                    return true;
-
-                case BinaryOpKind.LeqNumbers:
-                    invertedOp = BinaryOpKind.GeqNumbers;
-                    return true;
-
-                case BinaryOpKind.LeqDecimals:
-                    invertedOp = BinaryOpKind.GeqDecimals;
-                    return true;
-
-                case BinaryOpKind.LeqDateTime:
-                    invertedOp = BinaryOpKind.GeqDateTime;
-                    return true;
-
-                case BinaryOpKind.LeqDate:
-                    invertedOp = BinaryOpKind.GeqDate;
-                    return true;
-
-                case BinaryOpKind.LeqTime:
-                    invertedOp = BinaryOpKind.GeqTime;
-                    return true;
-
-                case BinaryOpKind.GtNumbers:
-                    invertedOp = BinaryOpKind.LtNumbers;
-                    return true;
-
-                case BinaryOpKind.GtDecimals:
-                    invertedOp = BinaryOpKind.LtDecimals;
-                    return true;
-
-                case BinaryOpKind.GtDateTime:
-                    invertedOp = BinaryOpKind.LtDateTime;
-                    return true;
-
-                case BinaryOpKind.GtDate:
-                    invertedOp = BinaryOpKind.LtDate;
-                    return true;
-
-                case BinaryOpKind.GtTime:
-                    invertedOp = BinaryOpKind.LtTime;
-                    return true;
-
-                case BinaryOpKind.GeqNumbers:
-                    invertedOp = BinaryOpKind.LeqNumbers;
-                    return true;
-
-                case BinaryOpKind.GeqDecimals:
-                    invertedOp = BinaryOpKind.LeqDecimals;
-                    return true;
-
-                case BinaryOpKind.GeqDateTime:
-                    invertedOp = BinaryOpKind.LeqDateTime;
-                    return true;
-
-                case BinaryOpKind.GeqDate:
-                    invertedOp = BinaryOpKind.LeqDate;
-                    return true;
-
-                case BinaryOpKind.GeqTime:
-                    invertedOp = BinaryOpKind.LeqTime;
-                    return true;
-
-                default:
-                    invertedOp = default;
-                    return false;
-            }
-        }
-
-        private static bool AllowedCoercions(UnaryOpNode unaryOp) =>
-            unaryOp.Op switch
-            {
-                UnaryOpKind.DateTimeToTime => true,
-                UnaryOpKind.DateToTime => true,
-                UnaryOpKind.TimeToDate => true,
-                UnaryOpKind.DateTimeToDate => true,
-                UnaryOpKind.TimeToDateTime => true,
-                UnaryOpKind.DateToDateTime => true,
-                _ => false
-            };
-
-        public bool TryGetFieldName(Context context, IntermediateNode node, out string fieldName)
-        {
-            IntermediateNode maybeScopeAccessNode;
-
-            // If the node had injected float coercion, then we need to pull scope access node from it.
-            if (node is CallNode functionCall &&
-                (functionCall.Function == BuiltinFunctionsCore.Float || functionCall.Function == BuiltinFunctionsCore.Value || functionCall.Function.Name == BuiltinFunctionsCore.IsBlank.Name) &&
-                functionCall.Args.Count == 1)
-            {
-                maybeScopeAccessNode = functionCall.Args[0];
-            }
-            else if (node is UnaryOpNode unaryOp && AllowedCoercions(unaryOp))
-            {
-                maybeScopeAccessNode = unaryOp.Child;
-            }
-            else
-            {
-                maybeScopeAccessNode = node;
-            }
-
-            if (maybeScopeAccessNode is ScopeAccessNode scopeAccessNode)
-            {
-                if (scopeAccessNode.Value is ScopeAccessSymbol scopeAccessSymbol)
-                {
-                    var callerScope = context.CallerNode.Scope;
-                    var callerId = callerScope.Id;
-                    if (scopeAccessSymbol.Parent.Id == callerId)
-                    {
-                        fieldName = scopeAccessSymbol.Name;
-
-                        if (fieldName == "Value" && ColumnMap.HasDistinct(context.CallerTableRetVal.ColumnMap))
-                        {
-                            fieldName = context.CallerTableRetVal.ColumnMap.Distinct;
-                        }
-
-                        return true;
-                    }
-                }
-            }
-
-            fieldName = default;
-            return false;
-        }
-    }
-}
