+++ src/PowerFx.Dataverse/Eval/Delegation/NodeVisitors/LazyEvalNodeVisitor.cs
@@ -1,0 +1,56 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using Microsoft.PowerFx.Core.IR;
+using Microsoft.PowerFx.Core.IR.Nodes;
+using Microsoft.PowerFx.Core.IR.Symbols;
+using Microsoft.PowerFx.Core.Localization;
+using Microsoft.PowerFx.Core.Texl;
+using Microsoft.PowerFx.Core.Types.Enums;
+using Microsoft.PowerFx.Core.Utils;
+using Microsoft.PowerFx.Dataverse.Eval.Core;
+using Microsoft.PowerFx.Dataverse.Eval.Delegation;
+using Microsoft.PowerFx.Dataverse.Eval.Delegation.DelegatedFunctions;
+using Microsoft.PowerFx.Types;
+using Microsoft.Xrm.Sdk.Metadata;
+using static Microsoft.PowerFx.Dataverse.DelegationEngineExtensions;
+using BinaryOpNode = Microsoft.PowerFx.Core.IR.Nodes.BinaryOpNode;
+using CallNode = Microsoft.PowerFx.Core.IR.Nodes.CallNode;
+using RecordNode = Microsoft.PowerFx.Core.IR.Nodes.RecordNode;
+using Span = Microsoft.PowerFx.Syntax.Span;
+using UnaryOpNode = Microsoft.PowerFx.Core.IR.Nodes.UnaryOpNode;
+
+namespace Microsoft.PowerFx.Dataverse
+{
+    internal partial class DelegationIRVisitor : RewritingIRVisitor<DelegationIRVisitor.RetVal, DelegationIRVisitor.Context>
+    {
+        public override RetVal Visit(LazyEvalNode node, Context context)
+        {
+            if (!context.IsPredicateEvalInProgress)
+            {
+                return base.Visit(node, context);
+            }
+
+            var child = node.Child.Accept(this, context);
+
+            if (child.IsDelegating)
+            {
+                return child;
+            }
+            else
+            {
+                if (!ReferenceEquals(child.OriginalNode, node.Child))
+                {
+                    node = new LazyEvalNode(node.IRContext, child.OriginalNode);
+                }
+
+                return Ret(node);
+            }
+        }
+    }
+}
