+++ src/PowerFx.Dataverse/Eval/Delegation/NodeVisitors/ResolvedObjectNodeVisitor.cs
@@ -1,0 +1,63 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using Microsoft.PowerFx.Core.IR;
+using Microsoft.PowerFx.Core.IR.Nodes;
+using Microsoft.PowerFx.Core.IR.Symbols;
+using Microsoft.PowerFx.Core.Localization;
+using Microsoft.PowerFx.Core.Texl;
+using Microsoft.PowerFx.Core.Types.Enums;
+using Microsoft.PowerFx.Core.Utils;
+using Microsoft.PowerFx.Dataverse.Eval.Core;
+using Microsoft.PowerFx.Dataverse.Eval.Delegation;
+using Microsoft.PowerFx.Dataverse.Eval.Delegation.DelegatedFunctions;
+using Microsoft.PowerFx.Types;
+using Microsoft.Xrm.Sdk.Metadata;
+using static Microsoft.PowerFx.Dataverse.DelegationEngineExtensions;
+using BinaryOpNode = Microsoft.PowerFx.Core.IR.Nodes.BinaryOpNode;
+using CallNode = Microsoft.PowerFx.Core.IR.Nodes.CallNode;
+using RecordNode = Microsoft.PowerFx.Core.IR.Nodes.RecordNode;
+using Span = Microsoft.PowerFx.Syntax.Span;
+using UnaryOpNode = Microsoft.PowerFx.Core.IR.Nodes.UnaryOpNode;
+
+namespace Microsoft.PowerFx.Dataverse
+{
+    // Rewrite the tree to inject delegation.
+    // If we encounter a dataverse table (something that should be delegated) during the walk, we either:
+    // - successfully delegate, which means rewriting to a call an efficient DelegatedFunction,
+    // - leave IR unchanged (don't delegate), but issue a warning.
+    internal partial class DelegationIRVisitor : RewritingIRVisitor<DelegationIRVisitor.RetVal, DelegationIRVisitor.Context>
+    {
+        // ResolvedObject is a symbol injected by the host.
+        // All Table references start as resolved objects.
+        public override RetVal Visit(ResolvedObjectNode node, Context context)
+        {
+            if (!context._ignoreDelegation && node.IRContext.ResultType is TableType aggType)
+            {
+                // Does the resolve object refer to a dataverse Table?
+                if (node.Value is NameSymbol nameSym)
+                {
+                    var symbolTable = nameSym.Owner;
+
+                    // We need to tell the difference between a direct table,
+                    // and another global variable that has that table's type (such as global := Filter(table, true).
+                    bool isRealTable = _hooks.IsDelegableSymbolTable(symbolTable);
+
+                    if (isRealTable)
+                    {
+                        var ret = new RetVal(_hooks, node, node, aggType, filter: null, orderBy: null, count: null, _maxRows, columnMap: null);
+                        return ret;
+                    }
+                }
+            }
+
+            // Just a regular variable, don't bother delegating.
+            return Ret(node);
+        }
+    }
+}
