+++ src/PowerFx.Dataverse/Eval/Delegation/Processors/ProcessFilter.cs
@@ -1,0 +1,73 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using Microsoft.PowerFx.Core.IR;
+using Microsoft.PowerFx.Core.IR.Nodes;
+using Microsoft.PowerFx.Core.IR.Symbols;
+using Microsoft.PowerFx.Core.Localization;
+using Microsoft.PowerFx.Core.Texl;
+using Microsoft.PowerFx.Core.Types.Enums;
+using Microsoft.PowerFx.Core.Utils;
+using Microsoft.PowerFx.Dataverse.Eval.Core;
+using Microsoft.PowerFx.Dataverse.Eval.Delegation;
+using Microsoft.PowerFx.Dataverse.Eval.Delegation.DelegatedFunctions;
+using Microsoft.PowerFx.Types;
+using Microsoft.Xrm.Sdk.Metadata;
+using static Microsoft.PowerFx.Dataverse.DelegationEngineExtensions;
+using BinaryOpNode = Microsoft.PowerFx.Core.IR.Nodes.BinaryOpNode;
+using CallNode = Microsoft.PowerFx.Core.IR.Nodes.CallNode;
+using RecordNode = Microsoft.PowerFx.Core.IR.Nodes.RecordNode;
+using Span = Microsoft.PowerFx.Syntax.Span;
+using UnaryOpNode = Microsoft.PowerFx.Core.IR.Nodes.UnaryOpNode;
+
+namespace Microsoft.PowerFx.Dataverse
+{
+    internal partial class DelegationIRVisitor : RewritingIRVisitor<DelegationIRVisitor.RetVal, DelegationIRVisitor.Context>
+    {
+        private RetVal ProcessFilter(CallNode node, RetVal tableArg, Context context)
+        {
+            if (node.Args.Count != 2)
+            {
+                return CreateNotSupportedErrorAndReturn(node, tableArg);
+            }
+
+            IntermediateNode predicate = node.Args[1];
+            IntermediateNode orderBy = tableArg.HasOrderBy ? tableArg.OrderBy : null;
+
+            var predicteContext = context.GetContextForPredicateEval(node, tableArg);
+            var pr = predicate.Accept(this, predicteContext);
+
+            if (!pr.IsDelegating)
+            {
+                // Though entire predicate is not delegable, pr._originalNode may still have delegation buried inside it.
+                if (!ReferenceEquals(pr.OriginalNode, predicate))
+                {
+                    node = new CallNode(node.IRContext, node.Function, node.Scope, new List<IntermediateNode>() { node.Args[0], pr.OriginalNode });
+                }
+
+                return CreateNotSupportedErrorAndReturn(node, tableArg);
+            }
+
+            RetVal result;
+
+            // If tableArg has top count, that means we need to materialize the tableArg and can't delegate.
+            if (tableArg.HasTopCount)
+            {
+                result = MaterializeTableAndAddWarning(tableArg, node);
+            }
+            else
+            {
+                // Since table was delegating it potentially has filter attached to it, so also add that filter to the new filter.
+                var filterCombined = tableArg.AddFilter(pr.Filter, node.Scope);
+                result = new RetVal(_hooks, node, tableArg._sourceTableIRNode, tableArg.TableType, filterCombined, orderBy: orderBy, count: null, _maxRows, tableArg.ColumnMap);
+            }
+
+            return result;
+        }
+    }
+}
